Speed %pyproject_save_files up significantly by using sets instead of lists
For packages with a lot of files, it took a really long time to run.
Profiling the code revealed most of the time is spent in PurePath.__eq__.

The code was using lists but checked
if to-be-added paths were not already in them.

Considering the order is insignificant
(the generated %files list is sorted at the end anyway),
the lists were essentially working as (very slow) sets.

Using sets instead of lists makes %pyproject_save_files over 20 times faster
(51.66 -> 2.39 seconds) for the ansible package (~62k files).

Checking if an item is in a list is O(N),
checking every added item is O(N**2).

Checking if an item is in a set is O(1),
checking every added item is O(N).

Additionally, with set, it is unnecessary to check for presence before addition,
so the code is easier.

(Commit message and removal of the check by Miro.)

Co-Authored-By: Miro Hrončok <miro@hroncok.cz>